fix: tests and add find_file function
I was having trouble running this project outside of PyCharm. Trying to
run this in VS Code or any other editor was a pain in the ass, and I
tried to run it once more to see if everything was fine. Tests were
passing on PyCharm, but they were not on VS Code, so I had to come up
with a way to fix everything. Ultimately, there was a problem with how
files were handled, so I had to come up with a way to find files
automatically. Hopefully, this won't affect performance that much. So I
fixed all the tests and added more tests to get this running, and
refactored every function that previously used hardcoded paths to find
files. I also moved main.py to the root dir, which fixes everything on
VS Code, it couldn't run previously because it couldn't find the src
module. Now everything should be fine here.